refactor: usage of auth and spotlight provider
