+++ apps/web/src/App.tsx
@@ -1,12 +1,10 @@
 import * as Sentry from '@sentry/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { HelmetProvider } from 'react-helmet-async';
-import { Route, Routes, Navigate, BrowserRouter, useLocation } from 'react-router-dom';
+import { Route, Routes, BrowserRouter } from 'react-router-dom';
 import { Integrations } from '@sentry/tracing';
-import decode from 'jwt-decode';
-import { IJwtPayload } from '@novu/shared';
-import { AuthProvider, useAuthContext } from './store/authContext';
-import { applyToken, getToken, getTokenPayload } from './store/useAuthController';
+import { AuthProvider } from './store/authContext';
+import { applyToken, getToken } from './store/useAuthController';
 import { ActivitiesPage } from './pages/activities/ActivitiesPage';
 import LoginPage from './pages/auth/LoginPage';
 import SignUpPage from './pages/auth/SignUpPage';
@@ -26,11 +24,8 @@ import { ENV, SENTRY_DSN, CONTEXT_PATH, LOGROCKET_ID } from './config';
 import { PromoteChangesPage } from './pages/changes/PromoteChangesPage';
 import { TemplateEditorProvider } from './components/templates/TemplateEditorProvider';
 import { TemplateFormProvider } from './components/templates/TemplateFormProvider';
-import { SpotLight } from './components/utils/Spotlight';
-import { SpotLightProvider } from './store/spotlightContext';
 import { LinkVercelProjectPage } from './pages/partner-integrations/LinkVercelProjectPage';
 import { ROUTES } from './constants/routes.enum';
-import { useBlueprint } from './hooks/useBlueprint';
 import { BrandPage } from './pages/brand/BrandPage';
 import { SegmentProvider } from './store/segment.context';
 import LogRocket from 'logrocket';
@@ -43,6 +38,7 @@ import { NotificationCenter } from './pages/quick-start/steps/NotificationCenter
 import { FrameworkSetup } from './pages/quick-start/steps/FrameworkSetup';
 import { Setup } from './pages/quick-start/steps/Setup';
 import { Trigger } from './pages/quick-start/steps/Trigger';
+import { RequiredAuth } from './components/layout/RequiredAuth';
 
 if (LOGROCKET_ID && window !== undefined) {
   LogRocket.init(LOGROCKET_ID, {
@@ -152,260 +148,80 @@ function App() {
         <BrowserRouter basename={CONTEXT_PATH}>
           <QueryClientProvider client={queryClient}>
             <AuthProvider>
-              <SpotLightProvider>
-                <Routes>
-                  <Route path={ROUTES.AUTH_SIGNUP} element={<SignUpPage />} />
-                  <Route path={ROUTES.AUTH_LOGIN} element={<LoginPage />} />
-                  <Route path={ROUTES.AUTH_RESET_REQUEST} element={<PasswordResetPage />} />
-                  <Route path={ROUTES.AUTH_RESET_TOKEN} element={<PasswordResetPage />} />
-                  <Route path={ROUTES.AUTH_INVITATION_TOKEN} element={<InvitationPage />} />
-                  <Route path={ROUTES.AUTH_APPLICATION} element={<CreateOrganizationPage />} />
+              <Routes>
+                <Route path={ROUTES.AUTH_SIGNUP} element={<SignUpPage />} />
+                <Route path={ROUTES.AUTH_LOGIN} element={<LoginPage />} />
+                <Route path={ROUTES.AUTH_RESET_REQUEST} element={<PasswordResetPage />} />
+                <Route path={ROUTES.AUTH_RESET_TOKEN} element={<PasswordResetPage />} />
+                <Route path={ROUTES.AUTH_INVITATION_TOKEN} element={<InvitationPage />} />
+                <Route path={ROUTES.AUTH_APPLICATION} element={<CreateOrganizationPage />} />
+                <Route
+                  path={ROUTES.PARTNER_INTEGRATIONS_VERCEL_LINK_PROJECTS}
+                  element={
+                    <RequiredAuth>
+                      <LinkVercelProjectPage type="create" />
+                    </RequiredAuth>
+                  }
+                />
+                <Route
+                  path={ROUTES.PARTNER_INTEGRATIONS_VERCEL_LINK_PROJECTS_EDIT}
+                  element={
+                    <RequiredAuth>
+                      <LinkVercelProjectPage type="edit" />
+                    </RequiredAuth>
+                  }
+                />
+                <Route element={<AppLayout />}>
+                  <Route path={ROUTES.ANY} element={<HomePage />} />
                   <Route
-                    path={ROUTES.PARTNER_INTEGRATIONS_VERCEL_LINK_PROJECTS}
+                    path={ROUTES.TEMPLATES_CREATE}
                     element={
-                      <RequiredAuth>
-                        <LinkVercelProjectPage type="create" />
-                      </RequiredAuth>
+                      <TemplateFormProvider>
+                        <TemplateEditorProvider>
+                          <TemplateEditorPage />
+                        </TemplateEditorProvider>
+                      </TemplateFormProvider>
                     }
                   />
                   <Route
-                    path={ROUTES.PARTNER_INTEGRATIONS_VERCEL_LINK_PROJECTS_EDIT}
+                    path={ROUTES.TEMPLATES_EDIT_TEMPLATEID}
                     element={
-                      <RequiredAuth>
-                        <LinkVercelProjectPage type="edit" />
-                      </RequiredAuth>
+                      <TemplateFormProvider>
+                        <TemplateEditorProvider>
+                          <TemplateEditorPage />
+                        </TemplateEditorProvider>
+                      </TemplateFormProvider>
                     }
                   />
-                  <Route element={<AppLayout />}>
-                    <Route
-                      path={ROUTES.ANY}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <HomePage />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.TEMPLATES_CREATE}
-                      element={
-                        <RequiredAuth>
-                          <TemplateFormProvider>
-                            <TemplateEditorProvider>
-                              <SpotLight>
-                                <TemplateEditorPage />
-                              </SpotLight>
-                            </TemplateEditorProvider>
-                          </TemplateFormProvider>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.TEMPLATES_EDIT_TEMPLATEID}
-                      element={
-                        <RequiredAuth>
-                          <TemplateFormProvider>
-                            <TemplateEditorProvider>
-                              <SpotLight>
-                                <TemplateEditorPage />
-                              </SpotLight>
-                            </TemplateEditorProvider>
-                          </TemplateFormProvider>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.TEMPLATES}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <NotificationList />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path="/general-started"
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <QuickStartWrapper>
-                              <GeneralStarter />
-                            </QuickStartWrapper>
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.QUICKSTART}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <Quickstart />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path="/quickstart/notification-center"
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <NotificationCenter />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path="/quickstart/notification-center/set-up"
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <FrameworkSetup />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path="/quickstart/notification-center/set-up/:framework"
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <Setup />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path="/quickstart/notification-center/trigger"
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <Trigger />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.ACTIVITIES}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <ActivitiesPage />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.SETTINGS}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <SettingsPage />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.INTEGRATIONS}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <IntegrationsStore />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.TEAM}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <MembersInvitePage />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.CHANGES}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <PromoteChangesPage />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path={ROUTES.SUBSCRIBERS}
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <SubscribersList />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                    <Route
-                      path="/brand"
-                      element={
-                        <RequiredAuth>
-                          <SpotLight>
-                            <BrandPage />
-                          </SpotLight>
-                        </RequiredAuth>
-                      }
-                    />
-                  </Route>
-                </Routes>
-              </SpotLightProvider>
+                  <Route path={ROUTES.TEMPLATES} element={<NotificationList />} />
+                  <Route
+                    path="/general-started"
+                    element={
+                      <QuickStartWrapper>
+                        <GeneralStarter />
+                      </QuickStartWrapper>
+                    }
+                  />
+                  <Route path={ROUTES.QUICKSTART} element={<Quickstart />} />
+                  <Route path="/quickstart/notification-center" element={<NotificationCenter />} />
+                  <Route path="/quickstart/notification-center/set-up" element={<FrameworkSetup />} />
+                  <Route path="/quickstart/notification-center/set-up/:framework" element={<Setup />} />
+                  <Route path="/quickstart/notification-center/trigger" element={<Trigger />} />
+                  <Route path={ROUTES.ACTIVITIES} element={<ActivitiesPage />} />
+                  <Route path={ROUTES.SETTINGS} element={<SettingsPage />} />
+                  <Route path={ROUTES.INTEGRATIONS} element={<IntegrationsStore />} />
+                  <Route path={ROUTES.TEAM} element={<MembersInvitePage />} />
+                  <Route path={ROUTES.CHANGES} element={<PromoteChangesPage />} />
+                  <Route path={ROUTES.SUBSCRIBERS} element={<SubscribersList />} />
+                  <Route path="/brand" element={<BrandPage />} />
+                </Route>
+              </Routes>
             </AuthProvider>
           </QueryClientProvider>
         </BrowserRouter>
       </HelmetProvider>
     </SegmentProvider>
   );
-}
-
-function jwtHasKey(key: string) {
-  const token = getToken();
-
-  if (!token) return false;
-  const jwt = decode<IJwtPayload>(token);
-
-  return jwt && jwt[key];
-}
-
-function RequiredAuth({ children }: any) {
-  useBlueprint();
-  const { logout } = useAuthContext();
-  const location = useLocation();
-
-  // TODO: remove after env migration
-  const payload = getTokenPayload();
-  if (payload && (payload as any).applicationId) {
-    logout();
-    window.location.reload();
-  }
-
-  // Logout if token.exp is in the past (expired)
-  if (payload && payload.exp && payload.exp <= Date.now() / 1000) {
-    logout();
-
-    return null;
-  }
-
-  if (!getToken()) {
-    return <Navigate to={ROUTES.AUTH_LOGIN} replace />;
-  } else if (
-    !jwtHasKey('organizationId') ||
-    (!jwtHasKey('environmentId') && location.pathname !== ROUTES.AUTH_APPLICATION)
-  ) {
-    return <Navigate to={ROUTES.AUTH_APPLICATION} replace />;
-  } else {
-    return children;
-  }
 }
 
 export default Sentry.withProfiler(App);
