+++ apps/web/src/components/layout/AppLayout.tsx
@@ -8,52 +8,61 @@ import { colors } from '../../design-system';
 import { IntercomProvider } from 'react-use-intercom';
 import { INTERCOM_APP_ID } from '../../config';
 import { HEADER_HEIGHT } from './constants';
+import { RequiredAuth } from './RequiredAuth';
+import { SpotLight } from '../utils/Spotlight';
+import { SpotLightProvider } from '../../store/spotlightContext';
 
 export function AppLayout() {
   return (
-    <ThemeProvider>
-      <SupportChatProvider>
-        <AppShell
-          padding="lg"
-          navbar={<SideNav />}
-          header={<HeaderNav />}
-          styles={(theme) => ({
-            root: { minHeight: '100vh', position: 'relative', zIndex: 1 },
-            body: {
-              minHeight: `calc(100vh - ${HEADER_HEIGHT})`,
-              '@media (max-width: 768px)': {
-                flexDirection: 'column',
-                height: 'auto',
+    <SpotLightProvider>
+      <ThemeProvider>
+        <SupportChatProvider>
+          <AppShell
+            padding="lg"
+            navbar={<SideNav />}
+            header={<HeaderNav />}
+            styles={(theme) => ({
+              root: { minHeight: '100vh', position: 'relative', zIndex: 1 },
+              body: {
+                minHeight: `calc(100vh - ${HEADER_HEIGHT})`,
+                '@media (max-width: 768px)': {
+                  flexDirection: 'column',
+                  height: 'auto',
+                },
               },
-            },
-            main: {
-              backgroundColor: theme.colorScheme === 'dark' ? colors.BGDark : colors.BGLight,
-              minHeight: 'auto',
-              padding: '30px',
-            },
-          })}
-        >
-          <Sentry.ErrorBoundary
-            fallback={({ error, resetError, eventId }) => (
-              <>
-                Sorry, but something went wrong. <br />
-                Our team been notified about it and we will look at it asap.
-                <br />
-                <code>
-                  <small style={{ color: 'lightGrey' }}>
-                    Event Id: {eventId}.
-                    <br />
-                    {error.toString()}
-                  </small>
-                </code>
-              </>
-            )}
+              main: {
+                backgroundColor: theme.colorScheme === 'dark' ? colors.BGDark : colors.BGLight,
+                minHeight: 'auto',
+                padding: '30px',
+              },
+            })}
           >
-            <Outlet />
-          </Sentry.ErrorBoundary>
-        </AppShell>
-      </SupportChatProvider>
-    </ThemeProvider>
+            <Sentry.ErrorBoundary
+              fallback={({ error, resetError, eventId }) => (
+                <>
+                  Sorry, but something went wrong. <br />
+                  Our team been notified about it and we will look at it asap.
+                  <br />
+                  <code>
+                    <small style={{ color: 'lightGrey' }}>
+                      Event Id: {eventId}.
+                      <br />
+                      {error.toString()}
+                    </small>
+                  </code>
+                </>
+              )}
+            >
+              <RequiredAuth>
+                <SpotLight>
+                  <Outlet />
+                </SpotLight>
+              </RequiredAuth>
+            </Sentry.ErrorBoundary>
+          </AppShell>
+        </SupportChatProvider>
+      </ThemeProvider>
+    </SpotLightProvider>
   );
 }
